refactor: re-structuring settings init functions
